+++ src/mkvtoolnix-gui/merge/mux_config.h
@@ -73,7 +73,7 @@ public:
 
   QString m_firstInputFileName;
   QString m_title, m_destination, m_destinationAuto, m_globalTags, m_segmentInfo, m_splitOptions;
-  QString m_segmentUIDs, m_previousSegmentUID, m_nextSegmentUID, m_chapters, m_chapterLanguage, m_chapterCharacterSet, m_chapterCueNameFormat, m_additionalOptions;
+  QString m_segmentUIDs, m_previousSegmentUID, m_nextSegmentUID, m_chapters, m_chapterLanguage, m_chapterCharacterSet, m_chapterCueNameFormat, m_chapterDelay, m_chapterStretchBy, m_additionalOptions;
   SplitMode m_splitMode;
   unsigned int m_splitMaxFiles;
   bool m_linkFiles, m_webmMode;
@@ -123,6 +123,7 @@ public:
   static void debugDumpSpecificTrackList(QList<Track *> const &tracks);
   static QString settingsType();
   static QString determineFirstInputFileName(QList<SourceFilePtr> const &files);
+  static QString formatDelayAndStretchBy(QString const &delay, QString const &stretchBy);
 };
 
 template<typename T>